clean up bot/mod.rs
